% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_dic.R
\name{add_user_word}
\alias{add_user_word}
\title{Add user word to dictionary}
\usage{
add_user_word(word, pos, score)
}
\arguments{
\item{word}{target word}

\item{pos}{pos}

\item{score}{score}
}
\description{
Add user word to dictionary
}
