% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{BuildOpt}
\alias{BuildOpt}
\title{Build options for Kiwi Class}
\format{
An object of class \code{EnumGenerator} of length 3.
}
\usage{
BuildOpt
}
\description{
Build options for Kiwi Class
}
\examples{
 BuildOpt
 BuildOpt$DEFAULT

}
\keyword{datasets}
