% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{cond_vowel}
\alias{cond_vowel}
\title{Condition Vowel}
\format{
Enum type defined on
}
\source{
\url{https://github.com/bab2min/Kiwi/blob/main/include/kiwi/Types.h#L180}
}
\usage{
cond_vowel
}
\description{
Condition Vowel
}
\keyword{datasets}
