% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{Dialect}
\alias{Dialect}
\title{Dialect Options}
\format{
An object of class \code{EnumGenerator} of length 11.
}
\usage{
Dialect
}
\description{
Dialect constants for analysis options.
}
\examples{
\dontrun{
Dialect
Dialect$STANDARD
}
}
\keyword{datasets}
