% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elbird-package.R
\docType{package}
\name{Elbird-package}
\alias{Elbird}
\alias{Elbird-package}
\title{Elbird: Blazing Fast Morphological Analyzer based on kiwi(Korean Intelligent Word Identifier)}
\description{
This is the R wrapper package Kiwi(Korean Intelligent Word Identifier), a blazing fast speed morphological analyzer for Korean. 
             It supports configuration of user dictionary and detection of unregistered nouns based on frequency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/Elbird}
  \item Report bugs at \url{https://github.com/mrchypark/Elbird/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
