% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{enums}
\alias{enums}
\alias{Tags}
\alias{Match}
\title{Elbird enums}
\format{
An object of class \code{EnumGenerator} of length 47.

An object of class \code{EnumGenerator} of length 13.
}
\usage{
Tags

Match
}
\description{
Elbird enums
}
\keyword{datasets}
\keyword{internal}
