% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_dic.R
\name{export_user_word_to_dictionary}
\alias{export_user_word_to_dictionary}
\title{Export user dictionary data}
\usage{
export_user_word_to_dictionary(path)
}
\arguments{
\item{path}{dictionary}
}
\description{
Export user dictionary data
}
