% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_add_words}
\alias{extract_add_words}
\title{Extract unregistered words from corpus provided user.}
\usage{
extract_add_words(
  reader,
  min_cnt = 10,
  max_word_len = 10,
  min_score = 0.25,
  pos_score = -3
)
}
\arguments{
\item{word}{target word}

\item{pos}{pos}

\item{score}{score}
}
\description{
Extract unregistered words from corpus provided user.
}
