% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwi_model.R
\name{get_model}
\alias{get_model}
\title{Get kiwi model.}
\source{
\url{https://github.com/bab2min/Kiwi/releases}
}
\usage{
get_model(
  version = "v0.11.0",
  size = "base",
  path = model_path(),
  force = FALSE
)
}
\arguments{
\item{version}{model version. default is "v0.11.0"}

\item{size}{small, base, large model. default is "base".}

\item{path}{path for model files. default is model_path().}

\item{fresh}{}
}
\description{
Get kiwi model.
}
