% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize model for function api}
\usage{
init(
  num_workers = 0,
  model_size = "base",
  integrate_allomorph = TRUE,
  load_default_dict = TRUE
)
}
\arguments{
\item{num_workers}{\code{int(optional)}: use multi-thread core number. default is 0 which means use all core.}

\item{model_size}{\code{char(optional)}: kiwi model select. default is "base". "small", "large" is available.}

\item{integrate_allomorph}{\code{bool(optional)}: default is TRUE.}

\item{load_default_dict}{\code{bool(optional)}: use defualt dictionary. default is TRUE.}
}
\description{
Initialize model for function api
}
