% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joiner.R
\name{Joiner}
\alias{Joiner}
\title{Joiner Class}
\description{
Joiner class provides methods to compose morphemes into text.
}
\examples{
\dontrun{
kw <- Kiwi$new()
joiner <- kw$create_joiner()
joiner$add("테스트", "NNG")
joiner$get()
}
}
