% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwi.R
\name{Kiwi_Class}
\alias{Kiwi_Class}
\title{Kiwi Class}
\description{
Kiwi class is provide method for korean mophological analyze result.
}
\examples{
\dontrun{
  el <- Elbird$new()
  el$analyze("test")
  el$tokenize("test")
  }

## ------------------------------------------------
## Method `Kiwi_Class$get_tidytext_func`
## ------------------------------------------------

\dontrun{
   kw <- Kiwi$new()
   tidytoken <- kw$get_tidytext_func()
   tidytoken("test")
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Kiwi-print}{\code{Kiwi_Class$print()}}
\item \href{#method-Kiwi-new}{\code{Kiwi_Class$new()}}
\item \href{#method-Kiwi-add_user_word}{\code{Kiwi_Class$add_user_word()}}
\item \href{#method-Kiwi-add_pre_analyzed_words}{\code{Kiwi_Class$add_pre_analyzed_words()}}
\item \href{#method-Kiwi-add_rules}{\code{Kiwi_Class$add_rules()}}
\item \href{#method-Kiwi-load_user_dictionarys}{\code{Kiwi_Class$load_user_dictionarys()}}
\item \href{#method-Kiwi-extract_words}{\code{Kiwi_Class$extract_words()}}
\item \href{#method-Kiwi-analyze}{\code{Kiwi_Class$analyze()}}
\item \href{#method-Kiwi-tokenize}{\code{Kiwi_Class$tokenize()}}
\item \href{#method-Kiwi-split_into_sents}{\code{Kiwi_Class$split_into_sents()}}
\item \href{#method-Kiwi-get_tidytext_func}{\code{Kiwi_Class$get_tidytext_func()}}
\item \href{#method-Kiwi-clone}{\code{Kiwi_Class$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-print"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-print}{}}}
\subsection{Method \code{print()}}{
print method for \code{Kiwi} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-new"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-new}{}}}
\subsection{Method \code{new()}}{
Create a kiwi instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$new(
  num_workers = 0,
  model_size = "base",
  integrate_allomorph = TRUE,
  load_default_dict = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-add_user_word"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-add_user_word}{}}}
\subsection{Method \code{add_user_word()}}{
add user word with pos and score
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$add_user_word(word, tag, score, orig_word = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{\code{char(required)}: target word to add.}

\item{\code{tag}}{\code{Tags(required)}: tag information about word.}

\item{\code{score}}{\code{num(required)}: score information about word.}

\item{\code{orig_word}}{\code{char(optional)}: origin word.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-add_pre_analyzed_words"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-add_pre_analyzed_words}{}}}
\subsection{Method \code{add_pre_analyzed_words()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$add_pre_analyzed_words(form, analyzed, score)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{\code{char(required)}: target word to add analyzed result.}

\item{\code{analyzed}}{\code{data.frame(required)}: analyzed result expected.}

\item{\code{score}}{\code{num(required)}: score information about pre analyzed result.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-add_rules"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-add_rules}{}}}
\subsection{Method \code{add_rules()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$add_rules(tag, pattern, replacement, score)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tag}}{\code{Tags(required)}: target tag to add rules.}

\item{\code{pattern}}{\code{char(required)}: regular expression.}

\item{\code{replacement}}{\code{char(required)}: replace text.}

\item{\code{score}}{\code{num(required)}: score information about rules.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-load_user_dictionarys"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-load_user_dictionarys}{}}}
\subsection{Method \code{load_user_dictionarys()}}{
add user dictionary using text file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$load_user_dictionarys(user_dict_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_dict_path}}{\code{char(required)}: path of user dictionary file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-extract_words"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-extract_words}{}}}
\subsection{Method \code{extract_words()}}{
Extract Noun word candidate from texts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$extract_words(
  input,
  min_cnt,
  max_word_len,
  min_score,
  pos_threshold,
  apply = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{\code{char(required)}: target text data}

\item{\code{min_cnt}}{\code{int(required)}: minimum count of word in text.}

\item{\code{max_word_len}}{\code{int(required)}: max word length.}

\item{\code{min_score}}{\code{num(required)}: minimum score.}

\item{\code{pos_threshold}}{\code{num(required)}: pos threashold.}

\item{\code{apply}}{\code{bool(optional)}: apply extracted word as user word dict.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-analyze"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-analyze}{}}}
\subsection{Method \code{analyze()}}{
Analyze text to token and tag results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$analyze(
  text,
  top_n = 3,
  match_option = Match$ALL,
  stopwords = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{\code{char(required)}: target text.}

\item{\code{top_n}}{\code{int(optional)}: number of result. Default is 3.}

\item{\code{match_option}}{match_option \code{\link{Match}}: use Match. Default is Match$ALL}

\item{\code{stopwords}}{stopwords option. Default is FALSE which is use nothing.
If \code{TRUE}, use embaded stopwords dictionany.
If \code{char}: path of dictionary txt file, use file.
If \code{\link{Stopwords}} class, use it.
If not valid value, work same as FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} of result.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-tokenize"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-tokenize}{}}}
\subsection{Method \code{tokenize()}}{
Analyze text to token and pos result just top 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$tokenize(
  text,
  match_option = Match$ALL,
  stopwords = FALSE,
  form = "tibble"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{\code{char(required)}: target text.}

\item{\code{match_option}}{match_option \code{\link{Match}}: use Match. Default is Match$ALL}

\item{\code{stopwords}}{stopwords option. Default is FALSE which is use nothing.
If \code{TRUE}, use embaded stopwords dictionany.
If \code{char}: path of dictionary txt file, use file.
If \code{\link{Stopwords}} class, use it.
If not valid value, work same as FALSE.}

\item{\code{form}}{\code{char(optional)}: return form. default is "tibble".
"list", "tidytext" is available.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-split_into_sents"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-split_into_sents}{}}}
\subsection{Method \code{split_into_sents()}}{
Some text may not split sentence by sentence.
split_into_sents works split sentences to sentence by sentence.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$split_into_sents(
  text,
  match_option = Match$ALL,
  return_tokens = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{\code{char(required)}: target text.}

\item{\code{match_option}}{match_option \code{\link{Match}}: use Match. Default is Match$ALL}

\item{\code{return_tokens}}{\code{bool(optional)}: add tokenized resault.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-get_tidytext_func"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-get_tidytext_func}{}}}
\subsection{Method \code{get_tidytext_func()}}{
set function to tidytext unnest_tokens.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$get_tidytext_func(match_option = Match$ALL, stopwords = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{match_option}}{match_option \code{\link{Match}}: use Match. Default is Match$ALL}

\item{\code{stopwords}}{stopwords option. Default is TRUE which is
to use embaded stopwords dictionary.
If FALSE, use not embaded stopwords dictionary.
If char: path of dictionary txt file, use file.
If \code{\link{Stopwords}} class, use it.
If not valid value, work same as FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{function}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
   kw <- Kiwi$new()
   tidytoken <- kw$get_tidytext_func()
   tidytoken("test")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kiwi-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Kiwi-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi_Class$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
