% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwi_model.R
\name{kiwi_model_path}
\alias{kiwi_model_path}
\title{A simple exported version of model_path
Returns the kiwi model path.}
\usage{
kiwi_model_path()
}
\description{
A simple exported version of model_path
Returns the kiwi model path.
}
