% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{Match}
\alias{Match}
\title{Analyze Match Options.}
\format{
An object of class \code{EnumGenerator} of length 13.
}
\usage{
Match
}
\description{
Analyze Match Options.
}
\examples{
 Match
 Match$ALL

}
\keyword{datasets}
