% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwi_model.R
\name{model_exists}
\alias{model_exists}
\title{Verifies if model files exists.}
\usage{
model_exists(size = "all")
}
\arguments{
\item{size}{model size. default is "all" which is true that all three models must be present.}
}
\value{
bool
}
\description{
Verifies if model files exists.
}
