% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwi_model.R
\name{model_home}
\alias{model_home}
\title{A simple exported version of \code{\link[=kiwi_model_path]{kiwi_model_path()}}
Returns the kiwi model path.}
\usage{
model_home()
}
\value{
path char.
}
\description{
TODO explain ELBIRD_MODEL_HOME
}
