% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwi_model.R
\name{model_is_set}
\alias{model_is_set}
\title{Verifies if model exist}
\usage{
model_is_set(size = "all")
}
\arguments{
\item{size}{model size. default is "all" which is true that all three models must be present.}
}
\value{
bool
}
\description{
Verifies if model exist
}
