% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiwi_model.R
\name{model_works}
\alias{model_works}
\title{Verifies if models work fine.}
\usage{
model_works(size = "all")
}
\arguments{
\item{size}{model size. default is "all" which is true that all three models must be present.}
}
\value{
\code{logical} model work or not.
}
\description{
Verifies if models work fine.
}
\examples{
\dontrun{
  get_model("small")
  model_works("small")
}
}
