% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphset.R
\name{Morphset}
\alias{Morphset}
\title{Morphset Class}
\description{
Morphset class provides methods for managing morpheme sets that can be used
to block specific morphemes from analysis results.
}
\examples{
\dontrun{
  kw <- Kiwi$new()
  morphset <- kw$create_morphset()
  morphset$add("테스트", "NNG")
  result <- kw$analyze("테스트 문장", blocklist = morphset)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Morphset-print}{\code{Morphset$print()}}
\item \href{#method-Morphset-new}{\code{Morphset$new()}}
\item \href{#method-Morphset-add}{\code{Morphset$add()}}
\item \href{#method-Morphset-add_multiple}{\code{Morphset$add_multiple()}}
\item \href{#method-Morphset-get_handle}{\code{Morphset$get_handle()}}
\item \href{#method-Morphset-size}{\code{Morphset$size()}}
\item \href{#method-Morphset-get_morphemes}{\code{Morphset$get_morphemes()}}
\item \href{#method-Morphset-clear}{\code{Morphset$clear()}}
\item \href{#method-Morphset-clone}{\code{Morphset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-print"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-print}{}}}
\subsection{Method \code{print()}}{
print method for \code{Morphset} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-new"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-new}{}}}
\subsection{Method \code{new()}}{
Create a morphset instance with a C++ handle.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$new(handle)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{handle}}{C++ handle for the morphset object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-add"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-add}{}}}
\subsection{Method \code{add()}}{
Add a morpheme to the morphset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$add(form, tag)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form}}{\code{char(required)}: morpheme form to add.}

\item{\code{tag}}{\code{char(required)}: POS tag for the morpheme.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical} indicating success
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-add_multiple"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-add_multiple}{}}}
\subsection{Method \code{add_multiple()}}{
Add multiple morphemes to the morphset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$add_multiple(forms, tags)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{forms}}{\code{character vector}: morpheme forms to add.}

\item{\code{tags}}{\code{character vector}: POS tags for the morphemes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{logical vector} indicating success for each morpheme
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-get_handle"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-get_handle}{}}}
\subsection{Method \code{get_handle()}}{
Get the internal C++ handle for this morphset.
This is used internally by the Kiwi class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$get_handle()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
C++ handle
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-size"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-size}{}}}
\subsection{Method \code{size()}}{
Get the number of morphemes in this morphset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{integer} number of morphemes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-get_morphemes"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-get_morphemes}{}}}
\subsection{Method \code{get_morphemes()}}{
Get a list of all morphemes in this morphset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$get_morphemes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of morphemes with form and tag
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-clear"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-clear}{}}}
\subsection{Method \code{clear()}}{
Clear all morphemes from this morphset.
Note: This creates a new morphset handle, so existing references may become invalid.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$clear()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Morphset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Morphset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Morphset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
