% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{pos_tag}
\alias{pos_tag}
\title{POS Tag}
\format{
Enum type defined on
}
\source{
\url{https://github.com/bab2min/Kiwi/blob/main/include/kiwi/Types.h#L152}
}
\usage{
pos_tag
}
\description{
POS Tag
}
\keyword{datasets}
