% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_into_sents.R
\name{split_into_sents}
\alias{split_into_sents}
\title{split_into_sents}
\usage{
split_into_sents(text, return_tokens = FALSE)
}
\arguments{
\item{text}{target text.}

\item{return_tokens}{add tokenized resault.}
}
\description{
split_into_sents
}
