% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{Stopwords}
\alias{Stopwords}
\title{Stop word Class}
\description{
Stop word Class

Stop word Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Stopwords$new()}}
\item \href{#method-add}{\code{Stopwords$add()}}
\item \href{#method-add_from_dict}{\code{Stopwords$add_from_dict()}}
\item \href{#method-remove}{\code{Stopwords$remove()}}
\item \href{#method-save_user_dict}{\code{Stopwords$save_user_dict()}}
\item \href{#method-get}{\code{Stopwords$get()}}
\item \href{#method-use}{\code{Stopwords$use()}}
\item \href{#method-lists}{\code{Stopwords$lists()}}
\item \href{#method-clone}{\code{Stopwords$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$new(use_system_dict = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$add(form, tag = "NNP")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_from_dict"></a>}}
\if{latex}{\out{\hypertarget{method-add_from_dict}{}}}
\subsection{Method \code{add_from_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$add_from_dict(path, dict_name = "user")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remove"></a>}}
\if{latex}{\out{\hypertarget{method-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$remove(form = NULL, tag = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_user_dict"></a>}}
\if{latex}{\out{\hypertarget{method-save_user_dict}{}}}
\subsection{Method \code{save_user_dict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$save_user_dict(path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$get()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-use"></a>}}
\if{latex}{\out{\hypertarget{method-use}{}}}
\subsection{Method \code{use()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$use()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lists"></a>}}
\if{latex}{\out{\hypertarget{method-lists}{}}}
\subsection{Method \code{lists()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$lists()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Stopwords$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
