% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swtokenizer.R
\name{SwTokenizer}
\alias{SwTokenizer}
\title{SwTokenizer Class}
\description{
SwTokenizer class provides encode/decode helpers for subword tokenization.
}
\examples{
\dontrun{
kw <- Kiwi$new()
swt <- kw$create_swtokenizer("tokenizer.json")
swt$encode("안녕하세요")
}
}
