% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{Tags}
\alias{Tags}
\title{Tag list}
\format{
An object of class \code{EnumGenerator} of length 47.
}
\source{
\url{https://github.com/bab2min/Kiwi}
}
\usage{
Tags
}
\description{
Tags contains tag list for elbird.
}
\examples{
 \dontrun{
  Tags
  Tags$nnp
 }
}
\keyword{datasets}
