% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize_tibble}
\alias{tokenize_tibble}
\alias{tokenize_tbl}
\title{Tokenize funtion whit return tibble form}
\usage{
tokenize_tibble(text, match_option = Match$ALL, stopwords = FALSE)

tokenize_tbl(text, match_option = Match$ALL, stopwords = FALSE)
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Please check \code{\link[=tokenize]{tokenize()}} docs.
}
