% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\docType{data}
\name{TypoSet}
\alias{TypoSet}
\title{Typo Set Options}
\format{
An object of class \code{EnumGenerator} of length 6.
}
\usage{
TypoSet
}
\description{
Typo correction set constants.
}
\examples{
\dontrun{
TypoSet
TypoSet$BASIC_TYPO_SET
}
}
\keyword{datasets}
