% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el.R
\name{analyze}
\alias{analyze}
\alias{analyze_tbl}
\alias{analyze_tidytext}
\alias{analyze_tt}
\title{analyze}
\usage{
analyze(text, top_n = 1)

analyze_tbl(text, top_n = 1)

analyze_tidytext(text, top_n = 1)

analyze_tt(text, top_n = 1)
}
\arguments{
\item{text}{target text.}

\item{top_n}{Number of result. default is 1.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
analyze
}
