% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbird-package.R
\docType{package}
\name{elbird-package}
\alias{elbird}
\alias{elbird-package}
\title{elbird: Blazing Fast Morphological Analyzer Based on Kiwi(Korean Intelligent Word Identifier)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This is the R wrapper package Kiwi(Korean Intelligent Word Identifier), a blazing fast speed morphological analyzer for Korean. It supports configuration of user dictionary and detection of unregistered nouns based on frequency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/elbird/}
  \item Report bugs at \url{https://github.com/mrchypark/elbird/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (\href{https://orcid.org/0000-0001-6474-2570}{ORCID})

}
\keyword{internal}
