% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_words}
\alias{extract_words}
\title{Extract unregistered words from corpus provided user.}
\usage{
extract_words(reader, min_cnt = 10, max_word_len = 10, min_score = 0.25)
}
\arguments{
\item{reader}{target word}

\item{min_cnt}{pos}

\item{max_word_len}{score}

\item{min_score}{score}
}
\description{
Extract unregistered words from corpus provided user.
}
