% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{Kiwi}
\alias{Kiwi}
\title{kiwi R6 class}
\description{
kiwi R6 class

kiwi R6 class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Kiwi$new()}}
\item \href{#method-tokenize}{\code{Kiwi$tokenize()}}
\item \href{#method-split_into_sents}{\code{Kiwi$split_into_sents()}}
\item \href{#method-extract_words}{\code{Kiwi$extract_words()}}
\item \href{#method-extract_add_words}{\code{Kiwi$extract_add_words()}}
\item \href{#method-add_user_words}{\code{Kiwi$add_user_words()}}
\item \href{#method-add_pre_analyzed_words}{\code{Kiwi$add_pre_analyzed_words()}}
\item \href{#method-save_user_dictionarys}{\code{Kiwi$save_user_dictionarys()}}
\item \href{#method-load_user_dictionarys}{\code{Kiwi$load_user_dictionarys()}}
\item \href{#method-clone}{\code{Kiwi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$new(
  num_workers = 0,
  model_size = "base",
  integrate_allomorph = TRUE,
  load_default_dict = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tokenize"></a>}}
\if{latex}{\out{\hypertarget{method-tokenize}{}}}
\subsection{Method \code{tokenize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$tokenize(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-split_into_sents"></a>}}
\if{latex}{\out{\hypertarget{method-split_into_sents}{}}}
\subsection{Method \code{split_into_sents()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$split_into_sents(text)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_words"></a>}}
\if{latex}{\out{\hypertarget{method-extract_words}{}}}
\subsection{Method \code{extract_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$extract_words(variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract_add_words"></a>}}
\if{latex}{\out{\hypertarget{method-extract_add_words}{}}}
\subsection{Method \code{extract_add_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$extract_add_words(variables)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_user_words"></a>}}
\if{latex}{\out{\hypertarget{method-add_user_words}{}}}
\subsection{Method \code{add_user_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$add_user_words(word, pos, score, orig_word)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_pre_analyzed_words"></a>}}
\if{latex}{\out{\hypertarget{method-add_pre_analyzed_words}{}}}
\subsection{Method \code{add_pre_analyzed_words()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$add_pre_analyzed_words(form, analyzed, score)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_user_dictionarys"></a>}}
\if{latex}{\out{\hypertarget{method-save_user_dictionarys}{}}}
\subsection{Method \code{save_user_dictionarys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$save_user_dictionarys(user_dict_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load_user_dictionarys"></a>}}
\if{latex}{\out{\hypertarget{method-load_user_dictionarys}{}}}
\subsection{Method \code{load_user_dictionarys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$load_user_dictionarys(user_dict_path)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Kiwi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
