% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_into_sents.R
\name{split_into_sents}
\alias{split_into_sents}
\title{Split Sentences}
\usage{
split_into_sents(text, return_tokens = FALSE)
}
\arguments{
\item{text}{target text.}

\item{return_tokens}{add tokenized resault.}
}
\description{
Some text may not split sentence by sentence.
split_into_sents works split sentences to sentence by sentence.
}
\examples{
\dontrun{
 split_into_sents("text")
 split_into_sents("text", return_tokens = TRUE)
}
}
