% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\alias{tokenize_tbl}
\alias{tokenize_tidytext}
\alias{tokenize_tt}
\alias{tokenize_tidy}
\title{tokenize}
\usage{
tokenize(text)

tokenize_tbl(text)

tokenize_tidytext(text)

tokenize_tt(text)

tokenize_tidy(text)
}
\arguments{
\item{text}{target text.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
tokenize
}
